[HashSet] Replaced the lists with hashSets for improved performances and also because it made more sense in all cases. That is becuase, for example a world state can contain the same relation only once, this is true with sets, not with lists, and also search operations are much quicker since they employ hashcodes
[getPossibleActions]
Changed the algorithm so that it doesn't rely anymore on list indexes.
Now it keeps a reference to the entity which is sobstituting, so that
the computed actions are always coherent

[applyAction]
Solved a bug arose since we are using hashsets which rely on hashcodes.
In order to obtain the resulting state from the actions we used to clone
the current one and then change the VALUE of the relations in order to
reflect the new changes of the world. This cannot be done with hashed
collections because then the value changes but not the hashcode
associated, so when searching the collections it outputs wrong results.
Now we remove the relations and add the new one, so a new hashcode
is computed