[Abstraction] when we are performing asbtraction we just need to parse the current state to the level of detail which we are now simulating. we can replace the current node with this parsed state without losing the history because it is kept into the lastObservedStates list.
[Refinement]
when we refine we must roll back history until the current node is root:
this does not mean that we are going back to the beginning, it just
means that we are going back to the last abstraction which sobstituted
the node with a new one that had no parent.
From there we use BFS to map each action to a list of equal actions
and re-roll forward history.